FIX DEAD-TIME: t_DTS from t_CK_INT (170 MHz), not timer_clk
Root cause: DT generator uses t_CK_INT (APB2 timer clock BEFORE PSC),
not timer_clk (after PSC). At 170 MHz: t_DTS = 5.88 ns, not 100 ns.
- Add: get_tim_ck_int(), encode/decode_dtg_ticks()
- Fix: PWM_DebugConfig dt=uint32_t (was uint8_t, truncated >255)
- Fix: PWM_Init dtg from t_CK_INT
- Fix: PWM_SetDeadTime_ns uses proper encoding
- Fix: GUI spinbox 0-3500 ns (was 0-255 ticks)
- Verified: dt=100->100ns, 500->500ns, 1500->1505ns, 3000->3011ns

+++ main.c
@@ -127,7 +127,7 @@ int main(void) {
             }
             else if(sscanf(linebuf, "p=%u,%u,%u,%u", &u1, &u2, &u3, &u4) >= 3) {
                 if(u4 == 0) { u4 = 0x3F; }
-                PWM_DebugConfig((uint16_t)u1, (uint16_t)u2, (uint8_t)u3, (uint8_t)u4);
+                PWM_DebugConfig((uint16_t)u1, (uint16_t)u2, u3, (uint8_t)u4);
                 UART_SendTelemetry("@PWM:OK:arr=%u:duty=%u:dt=%u\r\n> ", u1, u2, u3);
             }
             else if(linebuf[0] == '1' && linebuf[1] == '\0') {
